Don't depend on mozjs, webrender etc for building the style component
Logically, styles shouldn't depend on these components in the first place, but they're built anyway due to the dependency on malloc_size_of. Providing a bit more granularity might make more sense.

Additionally, currently mozjs depends on libz-sys which can't be built with emscripten; similarly, webrender_api depends on ipc-channel which can't be built with emscripten. It's unnecessary to criple the whole style component because of this.

Signed-off-by: Victor Porof <victor.porof@gmail.com>